update multiplier, ams, and fms parameter descriptions
