+++ src/MiniBoss.cpp
@@ -17,40 +17,11 @@
 #include "plugin.hpp"
 #include "dsp/triggers.hpp"
 #include "dsp/yamaha_ym2612/feedback_operator.hpp"
+#include "engine/yamaha_ym2612_params.hpp"
 
 // ---------------------------------------------------------------------------
 // MARK: Module
 // ---------------------------------------------------------------------------
-
-/// @brief A parameter quantity for the YM2612 LFO.
-struct LFOQuantity : rack::ParamQuantity {
-    /// @brief Return the value as a formatted string.
-    inline std::string getDisplayValueString() override {
-        const char* labels[8] = {
-            "3.98",
-            "5.56",
-            "6.02",
-            "6.37",
-            "6.88",
-            "9.63",
-            "48.1",
-            "72.2"
-        };
-        const int index = getValue();
-        if (index < 0 || index > 7) return "?";
-        return labels[index];
-    }
-};
-
-/// @brief A parameter quantity for the YM2612 Multiplier.
-struct MultiplierQuantity : rack::ParamQuantity {
-    /// @brief Return the value as a formatted string.
-    inline std::string getDisplayValueString() override {
-        const int value = getValue();
-        if (value == 0) return "1/2";
-        return std::to_string(value);
-    }
-};
 
 /// A Eurorack FM operator module based on the Yamaha YM2612.
 struct MiniBoss : rack::Module {
@@ -168,10 +139,10 @@ struct MiniBoss : rack::Module {
         configParam(PARAM_SL,  0,  15,  15, "Sustain Level");
         configParam(PARAM_D2,  0,  31,   0, "Sustain Rate");
         configParam(PARAM_RR,  0,  15,  15, "Release Rate");
-        configParam<MultiplierQuantity>(PARAM_MUL, 0,  15,   1, "Multiplier", "x");
+        configParam<MultiplierQuantity>(PARAM_MUL, 0, 15, 1);
         configParam(PARAM_RS,  0,   3,   0, "Rate Scaling");
-        configParam(PARAM_AMS, 0,   3,   0, "LFO amplitude modulation sensitivity");
-        configParam(PARAM_FMS, 0,   7,   0, "LFO frequency modulation sensitivity");
+        configParam<AMSQuantity>(PARAM_AMS, 0, 3, 0);
+        configParam<FMSQuantity>(PARAM_FMS, 0, 7, 0);
         configParam<BooleanParamQuantity>(PARAM_SSG_ENABLE, 0, 1, 0, "Looping Envelope");
         // reset the emulator
         onSampleRateChange();
